Safari-611.1.10/LayoutTests/js/codegen-jless-expected.txt

Tests particular unusual cases of jump-if-less codegen.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS !(true && undefined > 0) ? 'true' : 'false' is 'true'
PASS successfullyParsed is true

TEST COMPLETE
